+++ demisto_sdk/commands/common/content/objects/abstract_objects/json_object.py
@@ -3,7 +3,10 @@ from typing import Optional, Union
 from wcmatch.pathlib import Path
 
 import demisto_sdk.commands.common.content.errors as exc
-from demisto_sdk.commands.common.handlers import DEFAULT_JSON_HANDLER as json
+from demisto_sdk.commands.common.tools import (
+    get_file,
+    write_dict,
+)
 
 from .dictionary_based_object import DictionaryBasedObject
 
@@ -46,15 +49,14 @@ class JSONObject(DictionaryBasedObject):
     def _deserialize(self) -> None:
         """Load json to dictionary"""
         try:
-            self._as_dict = json.load(self._path.open())
+            self._as_dict = get_file(self._path, raise_on_error=True)
         except ValueError as e:
             raise exc.ContentSerializeError(self, self.path, str(e))
 
     def _serialize(self, dest_dir: Path):
         """Dump dictionary to json file"""
         dest_file = self._create_target_dump_dir(dest_dir) / self.normalize_file_name()
-        with open(dest_file, "w") as file:
-            json.dump(self.to_dict(), file)
+        write_dict(dest_file, data=self.to_dict())
         return [dest_file]
 
     def dump(self, dest_dir: Optional[Union[Path, str]] = None):