+++ demisto_sdk/commands/lint/linter.py
@@ -42,6 +42,7 @@ from demisto_sdk.commands.common.native_image import (
 from demisto_sdk.commands.common.timers import timer
 from demisto_sdk.commands.common.tools import (
     get_docker_images_from_yml,
+    get_file,
     get_id,
     get_pack_ignore_content,
     get_pack_name,
@@ -1281,8 +1282,8 @@ class Linter:
         )
         pack_metadata_file = pack_dir / PACKS_PACK_META_FILE_NAME
         logger.debug(f"Before reading content of {pack_metadata_file}")
-        with pack_metadata_file.open() as f:
-            pack_meta_content: Dict = json.load(f)
+        pack_meta_content: Dict = get_file(pack_metadata_file, raise_on_error=True)
+
         logger.debug(f"After reading content of {pack_metadata_file}")
         self._facts["support_level"] = pack_meta_content.get("support")
         if self._facts["support_level"] == "partner" and pack_meta_content.get(